Move Navigators under CustomQueries
Wings Navigators are specialized custom queries. This PR moves
Wings Navigators into services, and initializes them to be
used as standard custom queries.
