+++ lib/wings/hydra/pcdm/models/concerns/pcdm_valkyrie_behavior.rb
@@ -1,5 +1,4 @@
 require 'wings/active_fedora_converter'
-require 'wings/navigators/child_collections_navigator'
 
 module Wings
   module Pcdm
@@ -45,8 +44,7 @@ module Wings
       # @return [Enumerable<ActiveFedora::Base> | Enumerable<Valkyrie::Resource>] an enumerable over the child collections
       # @todo There is no guarantee to collection ordering until Hyrax is fully valkyrie-native, see issue 3784
       def child_collections(valkyrie: false)
-        @child_collections_navigator ||= ChildCollectionsNavigator.new(query_service: Hyrax.query_service)
-        resources = @child_collections_navigator.find_child_collections(resource: self)
+        resources = Hyrax.query_service.custom_queries.find_child_collections(resource: self)
         return resources if valkyrie
         resources.map { |r| Wings::ActiveFedoraConverter.new(resource: r).convert }
       end
